+++ Realm/RLMArray.h
@@ -73,6 +73,15 @@
 - (id)objectAtIndex:(NSUInteger)index;
 
 /**
+ Returns an NSArray of objects in the given range.
+
+ @param range   The range of objects to return.
+
+ @return An NSArray of objects in the given range.
+ */
+- (NSArray *)arrayWithRange:(NSRange)range;
+
+/**
  Returns the first object in the array.
  
  Returns `nil` if called on an empty RLMArray.