+++ Realm/RLMResults.mm
@@ -179,6 +179,14 @@ static inline void RLMResultsValidateInWriteTransaction(RLMResults *ar) {
     return RLMCreateObjectAccessor(_realm, _objectClassName, _backingView.get_source_ndx(index));
 }
 
+- (NSArray *)arrayWithRange:(NSRange)range {
+    NSMutableArray *array = [NSMutableArray arrayWithCapacity:range.length];
+    for (NSUInteger i = 0; i < range.length; i++) {
+        [array addObject:[self objectAtIndex:range.location + i]];
+    }
+    return array;
+}
+
 - (id)firstObject {
     if (self.count) {
         return [self objectAtIndex:0];